# openapi/openapi.json
{
  "components": {
    "parameters": {
      "AccessKeyId": {
        "in": "path",
        "name": "access_key_id",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "AdminLimit": {
        "in": "query",
        "name": "limit",
        "required": false,
        "schema": {
          "default": 100,
          "maximum": 200,
          "minimum": 1,
          "type": "integer"
        }
      },
...
    },
...
    "securitySchemes": {
      "CsrfToken": {
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "HmacAccessKey": {
        "description": "MediaHub-HMAC-SHA256 signed request with timestamp, nonce, body hash, and signed headers.",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      },
      "MetricsBearer": {
        "description": "Dedicated deployment metrics bearer token.",
        "scheme": "bearer",
        "type": "http"
      },
      "SessionCookie": {
        "in": "cookie",
        "name": "mediahub_session",
        "type": "apiKey"
      },
      "SignedMediaToken": {
        "description": "Short-lived token bound to one media revision and response policy.",
        "in": "query",
        "name": "token",
        "type": "apiKey"
      },
      "UploadCapability": {
        "description": "Short-lived capability bound to PUT and one upload session.",
        "in": "query",
        "name": "token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Docker Profile V1 API contract. Authenticated resource ownership is always derived from the session or API key, never from client-supplied owner IDs. Every response includes X-Request-Id; JSON errors additionally expose error.request_id.",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "MediaHub API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/access-keys/{access_key_id}": {
      "delete": {
        "parameters": [
          {
            "$ref": "#/components/parameters/AccessKeyId"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful response with no body"
          }
        },
        "security": [
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Revoke an access key"
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/components/parameters/AccessKeyId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccessKey"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessKey"
                }
              }
            },
            "description": "Successful response"
          }
        },
        "security": [
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Update an access key"
      }
    },
    "/api/v1/admin/applications": {
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/AdminLimit"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminApplication"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "List applications for system administration"
      }
    },
    "/api/v1/admin/applications/{application_id}/quota": {
      "patch": {
        "parameters": [
          {
            "$ref": "#/components/parameters/ApplicationId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateApplicationQuota"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminApplication"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Change an application storage quota"
      }
    },
    "/api/v1/admin/audit": {
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/AdminLimit"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminAudit"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "List deployment audit events"
      }
    },
    "/api/v1/admin/jobs": {
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/AdminLimit"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminJob"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "List jobs for system administration"
      }
    },
    "/api/v1/admin/settings": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSettings"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Read deployment download settings"
      },
      "patch": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSettings"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Update deployment download settings"
      }
    },
    "/api/v1/admin/storage": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStorage"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Read global storage totals"
      }
    },
    "/api/v1/admin/system/update": {
      "post": {
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemUpdateOperation"
                }
              }
            },
            "description": "Request accepted"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        },
        "security": [
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Request a controlled container image update"
      }
    },
...
    "/api/v1/admin/users": {
      "get": {
        "parameters": [
          {
            "$ref": "#/components/parameters/AdminLimit"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminUser"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "List users for system administration"
      }
    },
    "/api/v1/admin/users/{user_id}/status": {
      "patch": {
        "parameters": [
          {
            "$ref": "#/components/parameters/UserId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUpdateUserStatus"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUser"
                }
              }
            },
            "description": "Successful response"
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "CsrfToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Suspend or reactivate a user"
      }
    },
...
  }
}